Split crate_owner_invitations into MVC modules
